feat: 结果很清晰但有启发意义： 单行测试的结论： - 线性函数 parity: 63 个非零 w 中仅 1 个一致（w=63，全 1 行） - 三次/随机函数: 0 个一致 - Δ 检测完美匹配（0 误判），但一致性的门槛太高了 这意味着单行 M 候选对几乎所有非平凡函数都是不一致的——这不是检测方法的问题，而是单行变换本身几乎不可能恰好把 f 分割成两个常数值 fiber。 实际推论： 预筛选应该面向多行 M 中某行是否冗余，而不是单行是否有效。 比如对策略 2，如果已经选了 M = [w₁ ... w_{r-1}]，新行 w_r 要加进来时： - 解方程组 M[0..r-1]·Δ = 0 且 w_r·Δ = 1 - 如果无解 → w_r 的信息完全被前 r-1 行覆盖，跳过 - 如果有解 → w_r 可能贡献新信息 这比单行检测有用得多。需要我直接把这个做成  中的预筛逻辑吗？

+++ cpp/src/verify.cpp
@@ -4,7 +4,7 @@ bool verify_transform(const TruthTable& tt,
                       const std::vector<uint64_t>& g_tt,
                       int output_idx,
                       const uint32_t* M_rows,
-                      uint32_t b, int m, int n,
+                      uint64_t b, int m, int n,
                       int n_tests)
 {
     uint32_t M_offsets[64];
